edit backen

+++ backend/controllers/product.controller.js
@@ -54,11 +54,14 @@ export const updatedProduct = async (req, res) => {
 
 export const deleteProduct = async(req, res) => {
     const {id} = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid Product Id" });
+    }
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: "Product Deleted"});
     } catch (error) {
         console.error("Error in finding and deleting id ", error.message);
-        res.status(404).json( {success: false, message: "Product not found" });
+        res.status(500).json( {success: false, message: "Server error" });
     }
 }
